requirements.txt
# Core Polymarket dependencies
py-clob-client==0.18.0
web3==7.6.1
bip-utils==2.9.3
python-dotenv==1.0.0

# Speech recognition
vosk==0.3.45

# Media processing
yt-dlp==2023.10.13
ffmpeg-python==0.2.0
m3u8==3.5.0

# HTTP and API
requests==2.31.0
wget==3.2

# Data processing
numpy==1.24.3
pandas==2.0.3

# Error handling and retry logic
backoff==2.2.1

# Jupyter (for notebooks)
jupyter==1.0.0